chore(ios): KMP-enable remaining shared modules; move AppStartup (PR 4/N)
Adds iOS targets to the last four shared modules and moves `AppStartup`
from `:androidApp` into `:usecase`/commonMain — completing the
iOS-readiness of all KMP modules so PR 5 can wire the `NativeComponent`.

## iOS targets added (build.gradle.kts only)
- `:usecase`
- `:viewmodel`
- `:presentation-model`
- `:test-common`

Each: `applyDefaultHierarchyTemplate()` + iosX64/iosArm64/
iosSimulatorArm64. Zero source-code changes in `:usecase`,
`:presentation-model`, `:test-common`.

## `:viewmodel` portability fixes (2 ViewModels)
`System.currentTimeMillis()` is JVM-only — replaced with
`kotlinx.datetime.Clock.System.now().toEpochMilliseconds()` in:
- `HomeViewModel.submitButtonPress` (line 242)
- `FunctionListViewModel.openOrCloseDoor` (line 135)

Adds `kotlinx-datetime` dep to `:viewmodel` (same artifact PR 3 added
to `:data-local`).

## `AppStartup` move (androidApp → :usecase/commonMain)
`AppStartup.kt` was already pure KMP (constructor-injected
UseCases + Managers + DispatcherProvider + CoroutineScope, zero
Android imports) but lived under `androidApp/src/main/java/`. Moved
to `usecase/src/commonMain/kotlin/com/chriscartland/garage/usecase/`
so the iOS `NativeComponent` (PR 5) can expose it the same way
`AppComponent` does today.

Package: `com.chriscartland.garage` → `com.chriscartland.garage.usecase`.

Callers updated:
- `AppComponent.kt`: import line
- `AppStartupTest.kt`: added explicit import (test stays in
  `:androidApp` — it uses JUnit 4 + Mockito-free assertions, KMP-
  portable but no value to moving for this PR)

Detected via `git mv` rename.

## Validation
- All 47 validate.sh checks PASS (including `:usecase:testDebugUnitTest`
  + `:viewmodel:testDebugUnitTest` which exercise the changed VMs and
  `AppStartup`)
- Compiled `:usecase` + `:viewmodel` + `:presentation-model` +
  `:test-common` for all 3 iOS targets

## Merge order
PR 4 of N. Stacked on PR #822 (which is stacked on PR #821 → already
merged). GitHub will auto-rebase as each lands.

After PR 4, all shared KMP modules build for iOS. PR 5 adds SKIE
+ kotlin-inject KSP to `:iosFramework` and wires `NativeComponent`.
