+++ AndroidGarage/viewmodel/src/commonMain/kotlin/com/chriscartland/garage/viewmodel/FunctionListViewModel.kt
@@ -132,7 +132,9 @@ class DefaultFunctionListViewModel(
         viewModelScope.launch(dispatchers.io) {
             pushRemoteButtonUseCase(
                 buttonAckToken = ButtonAckToken.create(
-                    currentTimeMillis = System.currentTimeMillis(),
+                    currentTimeMillis = kotlinx.datetime.Clock.System
+                        .now()
+                        .toEpochMilliseconds(),
                     appVersion = appVersion,
                 ),
             )
